added a rates check when a transfer happens, changed DB to Postgresql

+++ banking/models.py
@@ -13,7 +13,7 @@ class User(AbstractUser):
         ('BTC', 'BTC'),
     )
     currency = models.CharField(max_length=3, choices=CURRENCY_CHOICES, blank=True)
-    balance = models.DecimalField(decimal_places=2, max_digits=10, blank=True, default=0)
+    balance = models.FloatField(blank=True, default=0)
     email = models.EmailField(unique=True)
     username = models.CharField(max_length=20)
 
@@ -26,7 +26,7 @@ class User(AbstractUser):
 
 class Transaction(models.Model):
     sender = models.ForeignKey(User, verbose_name='Sender', on_delete=models.CASCADE, related_name='sender')
-    amount = models.DecimalField(max_digits=10, decimal_places=2)
+    amount = models.FloatField()
     receiver = models.EmailField(verbose_name='Receiver', default=None)
     currency = models.CharField(verbose_name='Currency', max_length=3, blank=True)
     date = models.DateTimeField(verbose_name='Date', auto_now_add=True)
